添加删除节点功能

+++ onnx-modifier/server.py
@@ -376,23 +376,34 @@ def register_interface(app, request, send_file, temp_dir_path, init_file_path=No
 
     @app.route('/delete-custom-operator', methods=['POST'])
     def delete_custom_operator():
-        # 获取前端发送的数据
         data = request.json
-        operator_name = data.get('name')
-        operator_module = data.get('module')
+        operator_name = data['name']
+        operator_module = data['module']
+        operator_version = data.get('version', None)
 
-        # 打开 JSON 文件并删除指定算子
         with open('./static/onnx-metadata.json', 'r+', encoding='utf-8') as file:
             operators = json.load(file)
-            operators = [op for op in operators if
-                         not (op['name'] == operator_name and op['module'] == operator_module)]
 
-            # 重写文件
+            # 找到并删除匹配的算子
+            found = False
+            for op in operators:
+                if op['name'] == operator_name and op['module'] == operator_module \
+                        and op.get('version') == operator_version:
+                    operators.remove(op)
+                    found = True
+                    break
+
+            if not found:
+                return jsonify({"error": "Operator not found."}), 404
+
+            # 将更新后的算子列表写回文件
             file.seek(0)
             file.truncate()
             json.dump(operators, file, indent=4, ensure_ascii=False)
 
         return jsonify({"message": "Operator deleted successfully"})
+
+
     @app.route('/add-custom-operator', methods=['POST'])
     def add_custom_operator():
         # 获取前端发送的数据
@@ -417,8 +428,6 @@ def register_interface(app, request, send_file, temp_dir_path, init_file_path=No
             json.dump(file_data, file, indent=4, ensure_ascii=False)
 
         return jsonify({"message": "Custom operator added successfully"})
-
-
 
 
     @app.route('/init', methods=['POST'])
